price-backend/requirements.txt
beautifulsoup4==4.13.4
bidict==0.23.1
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.1.8
Flask==3.1.0
flask-cors==6.0.1
Flask-SocketIO==5.5.0
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
python-engineio==4.11.1
python-socketio==5.12.0
requests==2.32.4
simple-websocket==1.1.0
soupsieve==2.7
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
wsproto==1.2.0
gunicorn==23.0.0
python-dotenv==1.1.1
